Added workaround for R2014b/R2015a zero-width patch rendering issue
I discovered that a bug in HG2 versions of Matlab leads to odd patch rendering
when a patch is defined with no width (i.e. a line that traces back along itself,
as with a boundedline with e = 0).  Rather than rendering the edge line only, as
in previous versions, it instead draws a patch encompassing the convex hull of the
line.  I reported this bug to the Mathworks.  In the meantime, I've added a workaround
to the patch-calculation subfunction to avoid this behavior.