BUG: Disconnect outlinebounds on both ends
Previous behavior was to have the lower and upper bound lines
drawn together, in one fully connected line. This lead to a line
connecting the lower and upper bound lines at one edge of the plot
(the +ve end).

In this bug-fix, the two lines are still plotted at once as a single
object, but are disconnected by including a NaN between the lower
and upper line segments.